PlantBuddy_EdgeImpulse/server/requirements.txt
Flask>=2.0
pyserial>=3.4
requests>=2.0
